+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - The data loading indicator for the hydrograph now appears on top of the graph rather than with the date controls.
+- Cleaned up hydrograph rendering and made changes to make it possible to determine when rendering of the hydrograph is complete.
 
 ## [0.43.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.43.0...waterdataui-0.42.0)
 ### Added